[Fix] Consolidate connector shared API
