+++ packages/sdk/src/index.ts
@@ -76,11 +76,7 @@ export type {
     PageType,
     EditorResponse,
     SelectedLayoutFrame,
-    MetaData,
     Id,
-    ConnectorConfigOptions,
-    ConnectorConfigValue,
-    ConnectorConfigValueType,
 } from './types/CommonTypes';
 
 export type {
@@ -114,8 +110,6 @@ export {
     Scripting,
 } from './types/TextStyleTypes';
 export { ColorType } from './types/ColorStyleTypes';
-export * from './types/MediaConnectorTypes';
-export * from './types/FontConnectorTypes';
 export * from './types/ConnectorTypes';
 
 export { WellKnownConfigurationKeys } from './types/ConfigurationTypes';
