--- packages/sdk/src/types/FontConnectorTypes.ts
@@ -1,20 +1,0 @@
-import { Id } from './CommonTypes';
-
-export enum FontPreviewFormat {
-    Square = 'square',
-    Line = 'line',
-}
-
-export type FontFamily = {
-    id: Id;
-    name: string;
-    fontStylesCount: number;
-    extensions: string[];
-};
-
-export type FontStyle = {
-    id: Id;
-    name: string;
-    familyId: string;
-    familyName: string;
-};
