+++ packages/sdk/tsconfig.json
@@ -16,9 +16,10 @@
         "checkJs": true /* Report errors in .js files. Works in tandem with allowJs. */,
         // Ensure that .d.ts files are created by tsc, but not .js files
         "declaration": true,
-        "sourceMap": true
+        "sourceMap": true,
+        "paths": { "@connectorShared/*": ["../connectors/src/external/*"] }
     },
-    "include": ["src/**/*"],
+    "include": ["src/**/*", "../connectors/src/external/*"],
     "exclude": ["node_modules", "build"], // *** The files to not type check ***
     "typedocOptions": {
         "entryPoints": [
